Move fetch files to arbeidsgiver folder
